Implemented a filtering algorithm.
Added an ImageController to the main project. That controller will
process the HTTP POST request for filtering an image.
The new Filtering project will apply the filters to the given image and
will return the result as an ssv of "image strings".
Added a very fast Invert Filter that uses BitmapData and has an unsafe
block for direct memory access to the pixels of the image.