On drop the filtered image is shown.
Drag&drop works - shows the dropped image in place of the drop field.
The AJAX successfully sends an HTTP POST request.
The server applies the filter to the image and returns it.
The JS successfully shows the original and the filtered image.
Both images are now centered and the page rescales nicely.