Now handling sessions in a better way.
Now only logged users have sessions.
Users who are not logged do not have the privilage of a session.
The session contains the last dropped image.
So if a logged user goes to the About page, for example, and then
returns to the Home Page,  he will see his last dropped image and its
filtered image.

Also created a new file in the Filtering project - BitmapExtensions.
There i moved the methods that are extension methods to the Bitmap
class.
The ToImageString() method which converts a bitmap to a string for
easier HTTP transefer.
The Resize() method which resizes a Bitmap image to the given width and
height.