Added a custom Contrast Filter with two effects
If the Contrast Filter is applied with true then it also has a little
bit of a blur effect.
If the Contrast Filter is applied with false then the contrast of the
image is fixed and there is no blur, but a smooth result image.
Also rearanged the filters' order so that the more beautiful one are on
the top.